requirements.txt
aiohttp
APScheduler
asgiref
channels
channels_redis
cryptography
daphne
dj-database-url
Django
django-apscheduler
django-cleanup
django-cors-headers
django-extensions
django-oauth-toolkit
django-rest-framework
django-rest-knox
django-tinymce
djangorestframework
djangorestframework_simplejwt
drf-social-oauth2
folium
google-auth
google-genai
huggingface-hub
jwcrypto
# mysqlclient
numpy
oauthlib
opencage
openpyxl
pandas
paystackapi
paystackease
phonenumbers
pillow
pycryptodome
pydantic
pydantic_core
PyJWT
python-dateutil
python-decouple
redis
requests
requests-oauthlib
social-auth-app-django
social-auth-core
sqlparse
stripe
whitenoise
